Move AuraSaveStruct definition.
