+++ src/game/Database/CharacterDatabaseCache.h
@@ -20,6 +20,20 @@ struct PetSpellCache
 };
 typedef std::vector<PetSpellCache> PetSpells;
 
+struct AuraSaveStruct
+{
+    ObjectGuid casterGuid = 0;
+    uint32 itemLowGuid = 0;
+    uint32 spellId = 0;
+    uint32 stacks = 0;
+    uint32 charges = 0;
+    float  damage[MAX_EFFECT_INDEX] = { 0 };
+    uint32 periodicTime[MAX_EFFECT_INDEX] = { 0 };
+    int32 maxDuration = 0;
+    int32 duration = 0;
+    uint8 effIndexMask = 0;
+};
+
 // pet_aura
 typedef AuraSaveStruct PetAuraCache;
 typedef std::vector<PetAuraCache> PetAuras;
